js styles vs. InlineCSS
