+++ src/components/App.jsx
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react'
-import Viewer from './viewer/MirrorWrapper'
+import Viewer from './viewer/Viewer'
 import Header from './common/header/Menu'
 
 const App = React.createClass({
