+++ src/shared/styles.less
@@ -1,4 +1,6 @@
 @mainBodyColor: rgb(34, 34, 34);
+@number: 500px;
+
 
 body {
   background-color: @mainBodyColor;
